Refactored file names and Fixed import paths in TDDCycles module
